Fix issue with full reload after visiting dynamic page
